requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
python-multipart==0.0.9
pillow==10.2.0
httpx==0.26.0
python-dotenv==1.0.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
requests==2.31.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
